更新了pdf的修改
pdf文件可以正确修改

+++ code/App/Lib/Action/Admin/CourseAction.class.php
@@ -6,7 +6,24 @@
 ** 更新时间： 	2015/1/12
 ** 更新原因：	第一次写入
 */
-class CourseAction extends Action {
+class CourseAction extends CheckAction {
+	/* 属性名：		_needStatus
+	** 属性说明：	当前类需要的状态
+	*/
+	protected $_needStatus = 0;
+	
+	/* 方法名：		_initialize
+	** 方法说明：	所有类的构造函数
+	** 参数：		无
+	** 返回值：		无
+	*/
+    /* public function _initialize(){
+		//执行父类的构造函数
+		parent::_initialize();
+		if(!status_check($this->_needStatus)){
+			$this->error(status_error($_needStatus));
+		}
+	}*/
 	/* 方法名：		index
 	** 方法说明：	显示课程设计操作主页
 	** 参数：		无
@@ -24,9 +41,15 @@ class CourseAction extends Action {
 	** 返回值：		无
 	*/
     public function add(){
-		$System = D('System'); // 实例化Data数据模型	 
-		$this->system = $System->select();
-		$this->display();
+		//检测当前系统状态
+		$status = get_status();
+		if($status==$this->_needStatus){
+			$System = D('System'); // 实例化Data数据模型	 
+			$this->system = $System->select();
+			$this->display();
+		}else{
+			$this->error(status_error($this->_needStatus));
+		}
     }
 	
 	/* 方法名：		insert
@@ -35,17 +58,23 @@ class CourseAction extends Action {
 	** 返回值：		无
 	*/
 	public function insert(){
-		$Course = D('Course');
-        if($Course->create()) {
-			$result = $Course->add($Course->id);
-			if($result) {
-				$this->redirect('Admin/Course/index');
+		//检测当前系统状态
+		$status = get_status();
+		if($status==$this->_needStatus){
+			$Course = D('Course');
+			if($Course->create()) {
+				$result = $Course->add($Course->id);
+				if($result) {
+					$this->redirect('Admin/Course/index');
+				}else{
+					$this->error('写入错误！');
+				}
 			}else{
-				$this->error('写入错误！');
+				$this->error($Course->getError());
 			}
-        }else{
-            $this->error($Course->getError());
-        }
+		}else{
+			$this->error(status_error($this->_needStatus));
+		}
 	}
 	
 	/* 方法名：		edit
@@ -55,17 +84,23 @@ class CourseAction extends Action {
 	*/
 	public function edit($id = -1){
 		if($id == -1) $this->redirect('Admin/Course/index');//保护
-		$Course = D('Course');
-		$System = D('System');
-		$data=$Course->find($id);
-		if($data){
-			// 模板变量赋值
-			$this->system = $System->select();
-			$this->data = $data;
+		//检测当前系统状态
+		$status = get_status();
+		if($status==$this->_needStatus){
+			$Course = D('Course');
+			$System = D('System');
+			$data=$Course->find($id);
+			if($data){
+				// 模板变量赋值
+				$this->system = $System->select();
+				$this->data = $data;
+			}else{
+				$this->error('数据错误');
+			}
+			$this->display();
 		}else{
-			$this->error('数据错误');
+			$this->error(status_error($this->_needStatus));
 		}
-		$this->display();
 	}
 	
 	/* 方法名：		update
@@ -74,16 +109,22 @@ class CourseAction extends Action {
 	** 返回值：		无
 	*/
 	public function update(){
-		$Course = D('Course');
-		if($Course->create()) {
-			$result =   $Course->save();
-			if($result) {
-				$this->redirect('Admin/Course/index');
+		//检测当前系统状态
+		$status = get_status();
+		if($status==$this->_needStatus){
+			$Course = D('Course');
+			if($Course->create()) {
+				$result = $Course->save();
+				if($result) {
+					$this->redirect('Admin/Course/index');
+				}else{
+					$this->error('写入错误！');
+				}
 			}else{
-				$this->error('写入错误！');
+				$this->error($Course->getError());
 			}
 		}else{
-			$this->error($Course->getError());
+			$this->error(status_error($this->_needStatus));
 		}
 	}
 	
@@ -94,12 +135,18 @@ class CourseAction extends Action {
 	*/
 	public function delete($id = -1){
 		if($id == -1) $this->redirect('Admin/Course/index');//保护
-		$Course = D('Course');
-		$result=$Course->delete($id);
-		if($result) {
-			$this->redirect('Admin/Course/index');
+		//检测当前系统状态
+		$status = get_status();
+		if($status==$this->_needStatus){
+			$Course = D('Course');
+			$result = $Course->delete($id);
+			if($result) {
+				$this->redirect('Admin/Course/index');
+			}else{
+				$this->error('删除错误！');
+			}
 		}else{
-			$this->error('删除错误！');
+			$this->error(status_error($this->_needStatus));
 		}
 	}
